[235] refactorings on internal visitor, explorer and walker
