+++ plugins/org.eclipse.n4js.flowgraphs/src/org/eclipse/n4js/flowgraphs/analyses/EdgeGuideInternal.java
@@ -17,7 +17,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
-import org.eclipse.n4js.flowgraphs.analyses.PathExplorerInternal.PathWalkerInternal;
 import org.eclipse.n4js.flowgraphs.model.ControlFlowEdge;
 import org.eclipse.n4js.flowgraphs.model.JumpToken;
 import org.eclipse.n4js.flowgraphs.model.Node;
@@ -29,16 +28,16 @@ import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
 
 /**
- * The {@link EdgeGuideInternal} keeps track of all {@link PathWalkerInternal}s that are currently exploring a path on that
- * edge. In case an {@link EdgeGuideInternal} has no {@link PathWalkerInternal}s, it might be removed. When an edge of an
- * {@link EdgeGuideInternal} has more than one next edge, the {@link EdgeGuideInternal} will split up and all its
- * {@link PathWalkerInternal}s will fork. If there is only one next edge, the current edge of the {@link EdgeGuideInternal}
- * instance will be replaced.
+ * The {@link EdgeGuideInternal} keeps track of all {@link PathWalkerInternal}s that are currently exploring a path on
+ * that edge. In case an {@link EdgeGuideInternal} has no {@link PathWalkerInternal}s, it might be removed. When an edge
+ * of an {@link EdgeGuideInternal} has more than one next edge, the {@link EdgeGuideInternal} will split up and all its
+ * {@link PathWalkerInternal}s will fork. If there is only one next edge, the current edge of the
+ * {@link EdgeGuideInternal} instance will be replaced.
  * <p>
- * The {@link EdgeGuideInternal} keeps track of edges that enter or exit {@link FinallyBlock}s. The reason is, that these e.g.
- * entering edges will determine the correct exiting edges. (Consider that {@link FinallyBlock}s can be entered via a
- * {@link ReturnStatement} and will then exit directly to the next {@link FinallyBlock} or to the end of the method,
- * instead of executing statements that follow the {@link FinallyBlock}.)
+ * The {@link EdgeGuideInternal} keeps track of edges that enter or exit {@link FinallyBlock}s. The reason is, that
+ * these e.g. entering edges will determine the correct exiting edges. (Consider that {@link FinallyBlock}s can be
+ * entered via a {@link ReturnStatement} and will then exit directly to the next {@link FinallyBlock} or to the end of
+ * the method, instead of executing statements that follow the {@link FinallyBlock}.)
  */
 public class EdgeGuideInternal {
 	final NextEdgesProvider edgeProvider;
@@ -92,10 +91,10 @@ public class EdgeGuideInternal {
 	 * rules are implemented:
 	 * <ul>
 	 * <li/>If there exists no next edge with a context, then null is returned.
-	 * <li/>If there exists a next edge with a context, and the current {@link EdgeGuideInternal} instance has no context that
-	 * matches with one of the next edges, then all edges without context are returned.
-	 * <li/>If there exists a next edge with a context, and the current {@link EdgeGuideInternal} instance has a context that
-	 * matches with one of the next edges, the matching edge is returned.
+	 * <li/>If there exists a next edge with a context, and the current {@link EdgeGuideInternal} instance has no
+	 * context that matches with one of the next edges, then all edges without context are returned.
+	 * <li/>If there exists a next edge with a context, and the current {@link EdgeGuideInternal} instance has a context
+	 * that matches with one of the next edges, the matching edge is returned.
 	 * </ul>
 	 */
 	private List<ControlFlowEdge> findFinallyBlockContextEdge(List<ControlFlowEdge> nextEdges,
